+++ samplesheets/urls.py
@@ -48,12 +48,7 @@ urlpatterns = [
     ),
     # Javascript API views
     url(
-        regex=r'^irods/list/study/(?P<study>[0-9a-f-]+)$',
-        view=views.IrodsObjectListAPIView.as_view(),
-        name='irods_list',
-    ),
-    url(
-        regex=r'^irods/list/assay/(?P<assay>[0-9a-f-]+)$',
+        regex=r'^irods/list/assay/(?P<assay>[0-9a-f-]+)?path=(?P<path>[\w\-_/]+)$',
         view=views.IrodsObjectListAPIView.as_view(),
         name='irods_list',
     ),